[CQ] remove redundant cast (#8501)
No need for a cast here.

<img width="1866" height="154" alt="image" src="https://github.com/user-attachments/assets/7b865cd3-41be-4a93-8a43-25134ce33fa6" />

---

<details>
  <summary>Contribution guidelines:</summary><br>

- See our [contributor guide]([https://github.com/dart-lang/sdk/blob/main/CONTRIBUTING.md](https://github.com/flutter/flutter/blob/main/docs/contributing/Tree-hygiene.md#overview) for general expectations for PRs.
- Larger or significant changes should be discussed in an issue before creating a PR.
- Dart contributions to our repos should follow the [Dart style guide](https://dart.dev/guides/language/effective-dart) and use `dart format`.
- Java and Kotlin contributions should strive to follow Java and Kotlin best practices ([discussion](https://github.com/flutter/flutter-intellij/issues/8098)).
</details>

+++ src/io/flutter/test/DartTestEventsConverterZ.java
@@ -1,6 +1,10 @@
 package io.flutter.test;
 
-import com.google.gson.*;
+import com.google.gson.Gson;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonSyntaxException;
 import com.intellij.execution.testframework.TestConsoleProperties;
 import com.intellij.execution.testframework.sm.ServiceMessageBuilder;
 import com.intellij.execution.testframework.sm.runner.OutputToGeneralTestEventsConverter;
@@ -18,7 +22,11 @@ import org.jetbrains.annotations.Nullable;
 
 import java.lang.reflect.Type;
 import java.text.ParseException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -263,7 +271,7 @@ public class DartTestEventsConverterZ extends OutputToGeneralTestEventsConverter
     //if (test.getMetadata().skip) return true; // skipped tests are reported as ignored in handleTestStart(). testFinished signal must follow
 
     ServiceMessageBuilder testFinished = ServiceMessageBuilder.testFinished(test.getBaseName());
-    long duration = getTimestamp(obj) - (Long)myTestIdToTimestamp.get(test.getId());
+    long duration = getTimestamp(obj) - myTestIdToTimestamp.get(test.getId());
     testFinished.addAttribute("duration", Long.toString(duration));
 
     return finishMessage(testFinished, test.getId(), test.getValidParentId()) && checkGroupDone(test.getParent());
